addressing some comments from first review of pr2
